+++ src/runtime_src/core/common/api/xclbin_int.h
@@ -41,7 +41,7 @@ std::pair<const char*, size_t>
 get_axlf_section(const xrt::xclbin& xclbin, axlf_section_kind kind);
 
 // get_axlf_sections() - Retrieve specified sections
-std::vector< std::pair<const char*, size_t>>
+std::vector<std::pair<const char*, size_t>>
 get_axlf_sections(const xrt::xclbin& xclbin, axlf_section_kind kind);
 
 // read_xclbin() - Read specified xclbin file
