+++ src/runtime_src/core/include/experimental/CMakeLists.txt
@@ -12,6 +12,7 @@ set(XRT_EXPERIMENTAL_HEADER_SRC
   xrt_ini.h
   xrt_ip.h
   xrt_kernel.h
+  xrt_pskernel.h
   xrt_mailbox.h
   xrt_profile.h
   xrt_uuid.h
